save all files

+++ bot/common/constants.py
@@ -2,5 +2,4 @@ from typing import Final
 
 
 TEXT_MESSAGE_INTERVAL: Final[int] = 1
-
-SUBSCRIBE_DISCOUNT: Final[int] = 0.5
+SUBSCRIBE_DISCOUNT: Final[float] = 0.5
